# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UniTracker API",
        "description": "L'API UniTracker permet de suivre en temps réel les bus universitaires, planifier les trajets, et recevoir des notifications intelligentes. Elle facilite la gestion des transports pour les étudiants et les chauffeurs via des requêtes HTTP sécurisées.",
        "termsOfService": "https://unitracker.ucbc.cd/terms",
        "contact": {
            "name": "Équipe Support UniTracker",
            "email": "support@unitracker.cd"
        },
        "license": {
            "name": "UCBC / UniTracker",
            "url": "https://unitracker.ucbc.cd/license"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://unitracker.ucbc.cd",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/bus-locations": {
            "post": {
                "tags": [
                    "BusLocation"
                ],
                "summary": "Enregistrer la position GPS actuelle d’un bus",
                "operationId": "a2ab7dee6891bb3fbe0079c01ab98d9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bus_id",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "bus_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -1.6345623
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 29.1245891
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coordonnées enregistrées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "bus_id": {
                                                    "type": "integer"
                                                },
                                                "latitude": {
                                                    "type": "number"
                                                },
                                                "longitude": {
                                                    "type": "number"
                                                },
                                                "captured_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bus/{bus_id}/last-location": {
            "get": {
                "tags": [
                    "BusLocation"
                ],
                "summary": "Afficher la dernière position connue d’un bus",
                "operationId": "6b9c0adfea6da7cc8c0d551523b02594",
                "parameters": [
                    {
                        "name": "bus_id",
                        "in": "path",
                        "description": "ID du bus",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dernière position du bus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "bus_id": {
                                            "type": "integer"
                                        },
                                        "latitude": {
                                            "type": "number"
                                        },
                                        "longitude": {
                                            "type": "number"
                                        },
                                        "captured_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune position trouvée pour ce bus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
}